refactor: separate traffic light related code to another file
